[iOS] Update screen and container
